Fix signed/unsigned warning

+++ src/blobber.cc
@@ -155,7 +155,7 @@ void Foozle::Work_Slurp(uv_work_t* req) {
     if (gif_file->SColorMap) {
         int colors = clamp(0, gif_file->SColorMap->ColorCount, 256);
         
-        for (size_t i = 0; i < colors; i++) {
+        for (int i = 0; i < colors; i++) {
             GifColorType color = gif_file->SColorMap->Colors[i];
             foozle->palette[i] = color.Red | (color.Green<<8) | (color.Blue<<16) | 0xff000000;
         }
